backend/requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
psycopg2-binary==2.9.5
scikit-learn==1.2.2
pandas==2.0.0
numpy==1.24.2
gunicorn==20.1.0
python-dotenv==1.0.0